requirements.txt
numpy==1.16.1
tensorflow==1.15.0
flask==1.1.1
Pillow==6.2.0
gunicorn